move pollard util tests to new file and format changes
